Fix for #25
Sometimes, `this.select` might not start as a blank slate. In the broken example given in the issue page, `author.id` was already occupying `this.select`, and thus new fields were simply being appended. This ensures that if we have data in `this.select` that doesn't end in `,`, we add it manually