Start working on relation filtering without expand.
Here we identify the node that filters on something like posts/id, and sets the relation in the included so that data can be pulled back without expand.

There is currently a bug tho - if the filter comes before the expand, the expand doesn't set the selections correctly. This is due to us already having visited the table with the filter, and setting select = '' (hardcoded)